refactor:도메인 양방향 의존 제거, 다른 애그리게이트 간 직접참조 제거
