+++ src/main/java/kitchenpos/domain/menu/MenuProduct.java
@@ -1,7 +1,5 @@
 package kitchenpos.domain.menu;
 
-import kitchenpos.domain.Product;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -17,9 +15,9 @@ public class MenuProduct {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     private Long seq;
-    @ManyToOne
-    @JoinColumn(name = "menu_id")
-    private Menu menu;
+    //    @ManyToOne
+//    @JoinColumn(name = "menu_id")
+//    private Menu menu;
     @ManyToOne
     @JoinColumn(name = "product_id")
     private Product product;
@@ -30,11 +28,9 @@ public class MenuProduct {
     }
 
     public MenuProduct(final Long seq,
-                       final Menu menu,
                        final Product product,
                        final long quantity) {
         this.seq = seq;
-        this.menu = menu;
         this.product = product;
         this.quantity = quantity;
     }
@@ -43,9 +39,9 @@ public class MenuProduct {
         return seq;
     }
 
-    public Menu getMenu() {
-        return menu;
-    }
+//    public Menu getMenu() {
+//        return menu;
+//    }
 
     public Product getProduct() {
         return product;
