+++ src/main/java/kitchenpos/domain/order/OrderLineItem.java
@@ -7,8 +7,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 
 @Entity
 public class OrderLineItem {
@@ -16,12 +14,14 @@ public class OrderLineItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long seq;
-    @ManyToOne
-    @JoinColumn(name = "order_id")
-    private Order order;
-    @ManyToOne
-    @JoinColumn(name = "menu_id")
-    private Menu menu;
+    //    @ManyToOne
+//    @JoinColumn(name = "order_id")
+//    private Order order;
+//    @ManyToOne
+//    @JoinColumn(name = "menu_id")
+//    private Menu menu;
+    private Long menuId;
+    //이 validate는 validator가 menurepository를 가지도록 해서 검증하기?
     @Column
     private long quantity;
