+++ src/main/java/kitchenpos/domain/table/OrderTable.java
@@ -5,8 +5,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 import java.util.Objects;
 
 @Entity
@@ -15,9 +13,9 @@ public class OrderTable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @ManyToOne
-    @JoinColumn(name = "table_group_id")
-    private TableGroup tableGroup;
+    //    @ManyToOne
+//    @JoinColumn(name = "table_group_id")
+//    private TableGroup tableGroup;
     @Column
     private int numberOfGuests;
     @Column
