+++ src/test/java/kitchenpos/Fixture/Fixture.java
@@ -1,9 +1,9 @@
 package kitchenpos.Fixture;
 
-import kitchenpos.domain.Product;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menu.MenuGroup;
 import kitchenpos.domain.menu.MenuProduct;
+import kitchenpos.domain.menu.Product;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
 
@@ -13,7 +13,7 @@ import java.util.List;
 public abstract class Fixture {
     public static Menu menuFixture(String name, BigDecimal price, MenuGroup menuGroup, List<MenuProduct> menuProducts) {
         Menu menu = new Menu(name, price, menuGroup);
-        menu.setMenuProducts(menuProducts);
+        menu.registerMenuProducts(menuProducts);
         return menu;
     }
